split app.vue to componets
