+++ src/components/MyFooter.vue
@@ -1,0 +1,12 @@
+<template>
+  <v-footer padless>
+    <v-col class="text-center" cols="12">
+      <strong>@ {{ new Date().getFullYear() }} osaguild.com</strong>
+    </v-col>
+  </v-footer>
+</template>
+<script>
+export default {
+  name: "MyFooter",
+};
+</script>
